Refactor tests into separate file
